refactor: 콘서트 도메인명 이벤트로 수정
