+++ src/main/java/com/example/openticket/domain/event/Category.java
@@ -1,4 +1,4 @@
-package com.example.openticket.domain.concert;
+package com.example.openticket.domain.event;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
